A demonstration on how animation stops working if there is a viewless rib in the chain
This commit adds a new rib between `Parent` and `Child2` called `Child2Parent`
This rib does not have a view, and the only thing it does is display `Child2`. (In the real world this rib could contain some business logic that decides which child to display)
This scenario causes the transition animations to fail for `Child2`.